% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{app_server}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
app_server
}
