% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_fun.R
\name{cv_boxplot}
\alias{cv_boxplot}
\title{All plot functions}
\usage{
cv_boxplot(data, player, var, title = "", axe = "", legend = NULL,
  color = NULL)
}
\arguments{
\item{data}{data compare}

\item{player}{current player}

\item{var}{variables name}

\item{title}{plot title}

\item{axe}{axe title}

\item{legend}{legend}

\item{color}{color}
}
\description{
All plot functions
}
