% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_tmja.R
\name{display_tmja}
\alias{display_tmja}
\title{display_tmja}
\usage{
display_tmja(condition, map, group, color, data, n)
}
\arguments{
\item{condition}{c}

\item{map}{c}

\item{group}{c}

\item{color}{c}

\item{data}{c}

\item{n}{c}
}
\description{
display_tmja
}
