% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cartography.R
\name{mod_carto_ui}
\alias{mod_carto_ui}
\alias{mod_carto_server}
\title{mod_carto_ui and mod_carto_server}
\usage{
mod_carto_ui(id)

mod_carto_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
mod_carto_ui and mod_carto_server

mod_carto_server
}
