% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cluster.R
\name{mod_cluster_ui}
\alias{mod_cluster_ui}
\alias{mod_cluster_server}
\title{mod_cluster_ui and mod_cluster_server}
\usage{
mod_cluster_ui(id, fifa19_raw)

mod_cluster_server(input, output, session, fifa19_raw, fifa19_simple)
}
\arguments{
\item{id}{shiny id}

\item{fifa19_raw}{dataset with all skill attributes}

\item{input}{internal}

\item{output}{interlan}

\item{session}{internal}

\item{fifa19_simple}{dataset with simplified skill attributes}
}
\description{
UI and server of the Shiny module to make classification with the fifa19 dataset
}
