% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_explo.R
\name{mod_explo_ui}
\alias{mod_explo_ui}
\alias{mod_explo_server}
\title{mod_explo_ui and mod_explo_server}
\usage{
mod_explo_ui(id)

mod_explo_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
mod_explo_ui and mod_explo_server

mod_explo_server
}
