% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filtre_joueur.R
\name{mod_filtre_joueur_server}
\alias{mod_filtre_joueur_server}
\title{mod_filtre_joueur_server}
\usage{
mod_filtre_joueur_server(input, output, sessions)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{sessions}{internal}
}
\description{
mod_filtre_joueur_server
}
