% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filtre_joueur.R
\name{mod_filtre_joueur_ui}
\alias{mod_filtre_joueur_ui}
\title{mod_filtre_joueur_ui and mod_filtre_joueur_server}
\usage{
mod_filtre_joueur_ui(id, fifa2019)
}
\arguments{
\item{id}{shiny id}

\item{fifa2019}{dataset fifa2019}
}
\description{
mod_filtre_joueur_ui and mod_filtre_joueur_server
}
