% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_overview.R
\name{mod_overview_ui}
\alias{mod_overview_ui}
\alias{mod_overview_server}
\title{mod_overview_ui and mod_overview_server}
\usage{
mod_overview_ui(id)

mod_overview_server(input, output, session, fifa19_raw, fifa19_simple)
}
\arguments{
\item{id}{id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{fifa19_raw}{data}

\item{fifa19_simple}{data}
}
\description{
mod_overview_ui and mod_overview_server

mod_overview_server
}
