% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_radarchart.R
\name{mod_radarchart_ui}
\alias{mod_radarchart_ui}
\alias{mod_radarchart_server}
\title{mod_radarchart_ui and mod_radarchart_server}
\usage{
mod_radarchart_ui(id)

mod_radarchart_server(input, output, session, joueur, vector_of_skills,
  data = NULL)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{joueur}{data frame of 1 line containing the current selected player line}

\item{vector_of_skills}{list of 2 elements : title and features}

\item{data}{data to display with the radarchart}
}
\description{
Shiny module to make and display a radarchart
}
