% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_timeseries.R
\name{mod_timeseries_ui}
\alias{mod_timeseries_ui}
\alias{mod_timeseries_server}
\title{mod_timeseries_ui and mod_timeseries_server}
\usage{
mod_timeseries_ui(id)

mod_timeseries_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
Shiny module
}
