% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_fun.R
\name{plot_radar}
\alias{plot_radar}
\title{plot_radar}
\usage{
plot_radar(data_compare = NULL, player, vector_of_skills = NULL,
  data_skills = NULL, color = NULL, radarmean = FALSE)
}
\arguments{
\item{data_compare}{data}

\item{player}{current player}

\item{vector_of_skills}{skills}

\item{data_skills}{data}

\item{color}{color}

\item{radarmean}{bool}
}
\description{
plot_radar
}
