% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cluster.R
\name{mod_cluster_ui}
\alias{mod_cluster_ui}
\alias{mod_cluster_server}
\title{mod_cluster_ui}
\usage{
mod_cluster_ui(id, fifa19_raw)

mod_cluster_server(input, output, session, fifa19_raw, fifa19_simple)
}
\arguments{
\item{id}{shiny id}

\item{fifa19_raw}{raw}

\item{input}{internal}

\item{output}{interlan}

\item{session}{internal}

\item{fifa19_simple}{simple}
}
\description{
mod_cluster_ui

mod_cluster_server
}
