% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cv_joueur.R
\name{mod_cv_joueur_ui}
\alias{mod_cv_joueur_ui}
\alias{mod_cv_joueur_server}
\title{mod_cv_joueur_ui and mod_cv_joueur_server}
\usage{
mod_cv_joueur_ui(id, fifa19)

mod_cv_joueur_server(input, output, session, fifa19, fifa19_simple,
  filtre_inputs)
}
\arguments{
\item{id}{shiny id}

\item{fifa19}{dataset with all skill attributes}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{fifa19_simple}{dataset with simplified skill attributes}

\item{filtre_inputs}{inputs from mod_filtre_joueur module}
}
\description{
Shiny module to display a football player CV
}
