% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_overview.R
\name{mod_overview_ui}
\alias{mod_overview_ui}
\alias{mod_overview_server}
\title{mod_overview_ui and mod_overview_server}
\usage{
mod_overview_ui(id)

mod_overview_server(input, output, session, fifa19_raw, fifa19_simple)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{fifa19_raw}{dataset with all skill attributes}

\item{fifa19_simple}{dataset with simplified skill attributes}
}
\description{
Shiny module for fifa2019 dataset quick view and basic statistics
}
