% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_random_forest.R
\name{mod_random_forest_ui}
\alias{mod_random_forest_ui}
\alias{mod_random_forest_server}
\title{mod_random_forest_ui and mod_random_forest_server}
\usage{
mod_random_forest_ui(id, fifa19_raw)

mod_random_forest_server(input, output, session, data_complete,
  ratio_train)
}
\arguments{
\item{id}{shiny id}

\item{fifa19_raw}{dataset selected}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{data_complete}{selected dataset}

\item{ratio_train}{ratio to use for training the model}
}
\description{
Shiny module to compute a random forest modele on the dataset
}
